app/requirements.txt
# essential
Django<=5.0.0
django-extensions==3.2.3
django-betterforms==2.0.0
dj-config-url==0.1.1
django-filter==23.5
django-tables2==2.7.0
crispy-bootstrap5==2023.10
django-crispy-forms==2.1
django-dbbackup==4.0.2
# django-easy-pdf3 @ git+https://github.com/Toxic5698/django-easy-pdf3@c79a945ab7bb1003bfefa246089654b0cbdc59ce
django-weasyprint==2.2.2
boto3==1.34.1
botocore==1.34.1
django-environ==0.11.2
django-ipware==6.0.3
django-storages==1.14.2
django-model-utils==4.4.0
email-validator==2.1.0.post1
factory-boy==3.3.0
Faker==21.0.0
gunicorn==21.2.0
ipython==8.18.1
workalendar==17.0.0
pendulum==3.0.0
psycopg2==2.9.9
pypdfium2==4.25.0
sentry-sdk==1.39.1

# dev
selenium==4.16.0



# appnope==0.1.3
# arabic-reshaper==3.0.0
# asgiref==3.5.2
# asn1crypto==1.5.1
# asttokens==2.1.0
# attrs==22.2.0
# backcall==0.2.0
# Brotli==1.0.9
# certifi==2022.12.7
# cffi==1.15.1
# charset-normalizer==3.1.0
# click==8.1.3
# cryptography==39.0.2
# cssselect2==0.7.0
# decorator==5.1.1
# django-weasyprint==2.2.0
# dnspython==2.3.0
# exceptiongroup==1.1.0
# executing==1.2.0
# fonttools==4.38.0
# html5lib==1.1
# idna==3.4
# iniconfig==2.0.0
# jedi==0.18.1
# jmespath==1.0.1
# lxml==4.9.2
# MarkupSafe==2.1.2
# matplotlib-inline==0.1.6
# oscrypto==1.3.0
# packaging==23.0
# parso==0.8.3
# pexpect==4.8.0
# pickleshare==0.7.5
# Pillow==9.3.0
# pluggy==1.0.0
# prompt-toolkit==3.0.32
# ptyprocess==0.7.0
# pure-eval==0.2.2
# pycparser==2.21
# pydyf==0.5.0
# Pygments==2.13.0
# pyHanko==0.17.2
# pyhanko-certvalidator==0.20.1
# pypdf==3.5.1
# pyphen==0.13.1
# pypng==0.20220715.0
# pytest==7.2.1
# pytest-django==4.5.2
# python-bidi==0.4.2
# python-dateutil==2.8.2
# pytz==2022.7.1
# pytz-deprecation-shim==0.1.0.post0
# PyYAML==6.0
# qrcode==7.4.2
# reportlab==3.6.12
# requests==2.28.2
# s3transfer==0.6.1
# six==1.16.0
# sqlparse==0.4.3
# stack-data==0.6.1
# svglib==1.5.1
# tinycss2==1.2.1
# tomli==2.0.1
# traitlets==5.5.0
# typing_extensions==4.5.0
# tzdata==2022.7
# tzlocal==4.2
# uritools==4.0.1
# urllib3==1.26.14
# wcwidth==0.2.5
# weasyprint==57.1
# webencodings==0.5.1
# whitenoise==6.4.0
# xhtml2pdf==0.2.9
# zopfli==0.2.2
